requirements.txt
python-dotenv
fastapi
uvicorn
langchain
langchain_community
langchain-google-genai
python-json-logger
PyYAML
redis
pydantic-settings
langgraph_checkpoint_postgres
pyrate-limiter
tiktoken
psycopg[binary] 
asyncpg
sqlalchemy
alembic
langchain-tavily 
tavily-python
uvicorn
psycopg2